refactor dense/sparse modules
