+++ src/mmore/rag/model/__init__.py
@@ -1,4 +1,4 @@
-from .dense import load_dense_model
+from .dense import DenseModel
 from .sparse import load_sparse_model
 
-__all__ = ['load_dense_model', 'load_sparse_model']
+__all__ = ['']
